Refactor makeout.sh to use dynamic executable name
Refactor script to use dynamic executable name based on input argument. Update compilation commands to use the dynamic executable variable.

+++ trunk/install/makeout.sh
@@ -4,7 +4,9 @@ if [ "$1" = "" ]; then
    echo "Example: after compiled your standard program by 'gcc -o main main.cc' "
    echo "         TYPE ---> '$0 main' "
    echo "         will generate .out files for each .in files in "`pwd`
-   exit 1
+   EXEC="./Main"
+else
+   EXEC="./$1"
 fi
 if [ "$EUID" -eq 0 ]; then
     echo "Error: This script should not be run as root." >&2
@@ -16,17 +18,15 @@ if [ -f "Gen.py" ]; then
     echo "Gen.py finished."
 fi
 
-EXEC="./Main"
-
 if [ -f "Main.c" ]; then
-    gcc -Wall -static -DONLINE_JUDGE -lm -o Main Main.c
+    gcc -Wall -static -DONLINE_JUDGE -lm -o  "$EXEC"  Main.c
     echo "Compiled Main.c to Main using gcc with -Wall -static -DONLINE_JUDGE -lm"
 elif [ -f "Main.cc" ]; then
-    g++ -Wall -static -DONLINE_JUDGE -lm -o Main Main.cc
+    g++ -Wall -static -DONLINE_JUDGE -lm -o  "$EXEC"  Main.cc
     echo "Compiled Main.cc to Main using g++ with -Wall -static -DONLINE_JUDGE -lm"
 else
     echo "Error: Neither Main.c nor Main.cc found" >&2
-    EXEC="./$1"
+   
 fi
 
 if [ -f "interactor.cc" ]; then
@@ -39,10 +39,13 @@ fi
 for INFILE in `ls *.in`
 do
         OUTFILE=`basename -s .in $INFILE`.out
+    if [ -f "$EXEC" ]; then
         if $EXEC < $INFILE > $OUTFILE ; then
                 echo "make out for $INFILE -> $OUTFILE <br>"
         else
                 echo "make out for $INFILE .....failed"
         fi
+    fi
 done
+
